type:rename rename appearence_data to brightness_data
